+++ src/classes/scripts.ts
@@ -84,6 +84,7 @@ export class Scripts {
       queueKeys.id,
       queueKeys.delayed,
       queueKeys.priority,
+      queueKeys.completed,
       queueKeys.events,
       queueKeys.delay,
     ];
@@ -230,6 +231,7 @@ export class Scripts {
       opts.lockDuration,
       job.opts?.parent?.id,
       job.opts?.parent?.queue,
+      job.parentKey,
     ];
 
     return keys.concat(args);